Extracted Dates to Lib
