+++ backend/module/eCampCore/test/Data/PeriodTestData.php
@@ -8,7 +8,7 @@ use Doctrine\Persistence\ObjectManager;
 use eCamp\Core\Entity\Camp;
 use eCamp\Core\Entity\Day;
 use eCamp\Core\Entity\Period;
-use eCamp\Core\Types\DateUTC;
+use eCamp\Core\Types\DateUtc;
 
 class PeriodTestData extends AbstractFixture implements DependentFixtureInterface {
     public static $PERIOD1 = Period::class.':PERIOD1';
@@ -21,8 +21,8 @@ class PeriodTestData extends AbstractFixture implements DependentFixtureInterfac
         $period = new Period();
         $period->setCamp($camp);
         $period->setDescription('Period1');
-        $period->setStart(new DateUTC('2000-01-01'));
-        $period->setEnd(new DateUTC('2000-01-03'));
+        $period->setStart(new DateUtc('2000-01-01'));
+        $period->setEnd(new DateUtc('2000-01-03'));
 
         $days = $period->getDurationInDays();
         for ($idx = 0; $idx < $days; ++$idx) {
