+++ backend/module/eCampCore/test/Entity/ScheduleEntryTest.php
@@ -8,7 +8,7 @@ use eCamp\Core\Entity\ActivityType;
 use eCamp\Core\Entity\Camp;
 use eCamp\Core\Entity\Period;
 use eCamp\Core\Entity\ScheduleEntry;
-use eCamp\Core\Types\DateUTC;
+use eCamp\Core\Types\DateUtc;
 use eCamp\LibTest\PHPUnit\AbstractTestCase;
 
 /**
@@ -22,7 +22,7 @@ class ScheduleEntryTest extends AbstractTestCase {
 
         $camp = new Camp();
 
-        $start = new DateUTC();
+        $start = new DateUtc();
         $end = clone $start;
         $end->add(new \DateInterval('P7D'));
